requirements.txt
absl-py==2.0.0
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
deepface==0.0.79
easyocr==1.7.1
filelock==3.12.4
fire==0.5.0
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.1
ninja==1.11.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pyparsing==3.1.1
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.2
seaborn==0.12.2
sentry-sdk==1.31.0
shapely==2.0.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.18
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.183
urllib3==2.0.5
Werkzeug==2.3.7
wrapt==1.14.1
zipp==3.16.2
